feat: add some exception handling

+++ src/main/java/com/mastery/java/task/dao/EmployeeDao.java
@@ -2,7 +2,9 @@ package com.mastery.java.task.dao;
 
 import com.mastery.java.task.dto.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,7 +22,7 @@ public class EmployeeDao {
     public Employee getEmployee(int id) {
         Optional<Employee> employee = repository.findById(id);
         if (employee.isEmpty()){
-            throw new RuntimeException("Employee is not found with ID: " + id);
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employee is not found with ID: " + id);
         }
         return employee.get();
     }
@@ -30,10 +32,18 @@ public class EmployeeDao {
     }
 
     public void updateEmployee(int id, Employee employee){
-        repository.save(employee);
+        if (isEmployeeExist(id)) repository.save(employee);
     }
 
     public void deleteEmployee(int id){
-        repository.deleteById(id);
+        if (isEmployeeExist(id)) repository.deleteById(id);
     }
+
+    private Boolean isEmployeeExist(int id) {
+        if (repository.findById(id).isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Employee is not found with ID: " + id);
+        }
+        return true;
+    }
+
 }
